feat(events): Refactor event handling and improve error management in event pages
