+++ components/features/events/EventForm.tsx
@@ -21,6 +21,7 @@ import {
   type CreateEventInput,
 } from "@/lib/actions/events";
 import { createEventSchema, updateEventSchema } from "@/lib/actions/events";
+import { formatDateTimeLocal } from "@/lib/utils/date";
 
 interface EventFormProps {
   teamId: string;
@@ -128,16 +129,6 @@ export default function EventForm({
     } finally {
       setIsSubmitting(false);
     }
-  };
-
-  // Format date for datetime-local input
-  const formatDateTimeLocal = (date: Date) => {
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    const hours = String(date.getHours()).padStart(2, "0");
-    const minutes = String(date.getMinutes()).padStart(2, "0");
-    return `${year}-${month}-${day}T${hours}:${minutes}`;
   };
 
   return (
